10-autogen/requirements.txt
pyautogen>=0.2.0
pytest
